+++ Casks/maccrab.rb
@@ -1,6 +1,6 @@
 cask "maccrab" do
-  version "1.7.8"
-  sha256 "d471f836ecb96c365522956a3fd49673b5074b87ecfeac2804ded1d416fd6178"
+  version "1.7.9"
+  sha256 "3fb9fe2e419af781b332a11db2d50f9fa48d82788d5b908a0f25a72604413e2d"
 
   url "https://github.com/peterhanily/maccrab/releases/download/v#{version}/MacCrab-v#{version}.dmg"
   name "MacCrab"